+++ matminer/featurizers/base.py
@@ -1,3 +1,5 @@
+from __future__ import division
+
 import numpy as np
 from six import string_types