+++ matminer/featurizers/tests/test_bandstructure.py
@@ -5,10 +5,12 @@ import os
 import pandas as pd
 import unittest
 
-from matminer.featurizers.bandstructure import BandFeaturizer, BranchPointEnergy
+from matminer.featurizers.bandstructure import BandFeaturizer, \
+    BranchPointEnergy, DOSFeaturizer
 from pymatgen import Structure
 from pymatgen.electronic_structure.bandstructure import BandStructureSymmLine, \
     BandStructure
+from pymatgen.electronic_structure.dos import CompleteDos
 from pymatgen.util.testing import PymatgenTest
 
 test_dir = os.path.join(os.path.dirname(__file__))
@@ -43,8 +45,33 @@ class BandstructureFeaturesTest(PymatgenTest):
         self.assertAlmostEqual(df_bf['direct_gap'][0], 2.557, 3)
         self.assertAlmostEqual(df_bf['n_ex1_norm'][0], 0.58413, 5)
         self.assertAlmostEqual(df_bf['p_ex1_norm'][0], 0.0, 5)
-        self.assertEquals(df_bf['is_gap_direct'][0], False)
+        self.assertEqual(df_bf['is_gap_direct'][0], False)
+        self.assertEqual(df_bf['n_ex1_degen'][0], 6)
+        self.assertEqual(df_bf['p_ex1_degen'][0], 1)
 
+class DOSFeaturesTest(PymatgenTest):
+
+    def setUp(self):
+        with open(os.path.join(test_dir, 'si_dos.json'), 'r') as sDOS:
+            si_dos = CompleteDos.from_dict(json.load(sDOS))
+        self.df = pd.DataFrame({'dos': [si_dos]})
+
+    def test_DOSFeaturizer(self):
+        df_df = DOSFeaturizer().featurize_dataframe(self.df,
+                                                    col_id=['dos'])
+
+        self.assertAlmostEqual(df_df['cbm_percents'][0][0], 0.258, 3)
+        self.assertAlmostEqual(df_df['cbm_locations'][0], [[0.0, 0.0, 0.0]])
+        self.assertEqual(df_df['cbm_species'][0], ['Si'])
+        self.assertEqual(df_df['cbm_characters'][0], ['s'])
+        self.assertEqual(df_df['cbm_coordinations'][0], ['tet'])
+        self.assertEqual(df_df['cbm_significant_contributors'][0], 4)
+        self.assertAlmostEqual(df_df['vbm_percents'][0][0], 0.490, 3)
+        self.assertEqual(df_df['vbm_locations'][0], [[0.0, 0.0, 0.0]])
+        self.assertEqual(df_df['vbm_species'][0], ['Si'])
+        self.assertEqual(df_df['vbm_characters'][0], ['p'])
+        self.assertEqual(df_df['vbm_coordinations'][0], ['tet'])
+        self.assertEqual(df_df['vbm_significant_contributors'][0], 2)
 
 if __name__ == '__main__':
     unittest.main()