+++ matminer/featurizers/tests/test_structure.py
@@ -2,17 +2,15 @@
 # Copyright (c) Pymatgen Development Team.
 # Distributed under the terms of the MIT License.
 
-from __future__ import unicode_literals
+from __future__ import unicode_literals, division
 
 import unittest
 
 import numpy as np
-from math import fabs
 
 from pymatgen import Structure, Lattice, Molecule
 from pymatgen.util.testing import PymatgenTest
 
-from matminer.featurizers.site import OPSiteFingerprint
 from matminer.featurizers.structure import DensityFeatures, \
     RadialDistributionFunction, \
     RadialDistributionFunctionPeaks, PartialRadialDistributionFunction, \
@@ -80,13 +78,13 @@ class StructureFeaturesTest(PymatgenTest):
         rdf = rdforig[0]
 
         # Make sure it the last bin is cutoff-bin_max
-        self.assertAlmostEquals(max(rdf['distances']), 19.9)
+        self.assertAlmostEqual(max(rdf['distances']), 19.9)
 
         # Verify bin sizes
-        self.assertEquals(len(rdf['distribution']), 200)
+        self.assertEqual(len(rdf['distribution']), 200)
 
         # Make sure it gets all of the peaks
-        self.assertEquals(np.count_nonzero(rdf['distribution']), 116)
+        self.assertEqual(np.count_nonzero(rdf['distribution']), 116)
 
         # Check the values for a few individual peaks
         self.assertAlmostEqual(
@@ -98,16 +96,16 @@ class StructureFeaturesTest(PymatgenTest):
 
         # Make sure it finds the locations of non-zero peaks correctly
         peaks = RadialDistributionFunctionPeaks().featurize(rdforig)[0]
-        self.assertEquals(len(peaks), 2)
-        self.assertAlmostEquals(2.5, peaks[0])
-        self.assertAlmostEquals(1.5, peaks[1])
+        self.assertEqual(len(peaks), 2)
+        self.assertAlmostEqual(2.5, peaks[0])
+        self.assertAlmostEqual(1.5, peaks[1])
 
         # Repeat test with NaCl (omitting comments). Altering cutoff distance
         rdforig = RadialDistributionFunction(cutoff=10).featurize(self.nacl)
         rdf = rdforig[0]
-        self.assertAlmostEquals(max(rdf['distances']), 9.9)
-        self.assertEquals(len(rdf['distribution']), 100)
-        self.assertEquals(np.count_nonzero(rdf['distribution']), 11)
+        self.assertAlmostEqual(max(rdf['distances']), 9.9)
+        self.assertEqual(len(rdf['distribution']), 100)
+        self.assertEqual(np.count_nonzero(rdf['distribution']), 11)
         self.assertAlmostEqual(
             rdf['distribution'][int(round(2.8 / 0.1))], 27.09214168)
         self.assertAlmostEqual(
@@ -116,17 +114,17 @@ class StructureFeaturesTest(PymatgenTest):
             rdf['distribution'][int(round(9.8 / 0.1))], 3.024406467)
 
         peaks = RadialDistributionFunctionPeaks().featurize(rdforig)[0]
-        self.assertEquals(len(peaks), 2)
-        self.assertAlmostEquals(2.8, peaks[0])
-        self.assertAlmostEquals(4.0, peaks[1])
+        self.assertEqual(len(peaks), 2)
+        self.assertAlmostEqual(2.8, peaks[0])
+        self.assertAlmostEqual(4.0, peaks[1])
 
         # Repeat test with CsCl. Altering cutoff distance and bin_size
         rdforig = RadialDistributionFunction(
             cutoff=8, bin_size=0.5).featurize(self.cscl)
         rdf = rdforig[0]
-        self.assertAlmostEquals(max(rdf['distances']), 7.5)
-        self.assertEquals(len(rdf['distribution']), 16)
-        self.assertEquals(np.count_nonzero(rdf['distribution']), 5)
+        self.assertAlmostEqual(max(rdf['distances']), 7.5)
+        self.assertEqual(len(rdf['distribution']), 16)
+        self.assertEqual(np.count_nonzero(rdf['distribution']), 5)
         self.assertAlmostEqual(
             rdf['distribution'][int(round(3.5 / 0.5))], 6.741265585)
         self.assertAlmostEqual(
@@ -136,17 +134,17 @@ class StructureFeaturesTest(PymatgenTest):
 
         peaks = RadialDistributionFunctionPeaks(n_peaks=3).featurize(
             rdforig)[0]
-        self.assertEquals(len(peaks), 3)
-        self.assertAlmostEquals(3.5, peaks[0])
-        self.assertAlmostEquals(6.5, peaks[1])
-        self.assertAlmostEquals(5, 5, peaks[2])
+        self.assertEqual(len(peaks), 3)
+        self.assertAlmostEqual(3.5, peaks[0])
+        self.assertAlmostEqual(6.5, peaks[1])
+        self.assertAlmostEqual(5, 5, peaks[2])
 
     def test_prdf(self):
         # Test a few peaks in diamond
         # These expected numbers were derived by performing
         # the calculation in another code
         prdf = PartialRadialDistributionFunction().featurize(self.diamond)[0]
-        self.assertEquals(len(prdf.values()), 1)
+        self.assertEqual(len(prdf.values()), 1)
         self.assertAlmostEqual(
             prdf[('C', 'C')]['distribution'][int(round(1.4 / 0.1))], 0)
         self.assertAlmostEqual(
@@ -158,24 +156,24 @@ class StructureFeaturesTest(PymatgenTest):
         # Test a few peaks in CsCl, make sure it gets all types correctly
         prdf = PartialRadialDistributionFunction(cutoff=10).featurize(
             self.cscl)[0]
-        self.assertEquals(len(prdf.values()), 4)
+        self.assertEqual(len(prdf.values()), 4)
         self.assertAlmostEqual(max(prdf[('Cs', 'Cl')]['distances']), 10.0)
-        self.assertAlmostEquals(
+        self.assertAlmostEqual(
             prdf[('Cs', 'Cl')]['distribution'][int(round(3.6 / 0.1))], 0.477823197)
-        self.assertAlmostEquals(
+        self.assertAlmostEqual(
             prdf[('Cl', 'Cs')]['distribution'][int(round(3.6 / 0.1))], 0.477823197)
-        self.assertAlmostEquals(
+        self.assertAlmostEqual(
             prdf[('Cs', 'Cs')]['distribution'][int(round(3.6 / 0.1))], 0)
 
         # Do Ni3Al, make sure it captures the antisymmetry of Ni/Al sites
         prdf = PartialRadialDistributionFunction(
             cutoff=10, bin_size=0.5).featurize(self.ni3al)[0]
-        self.assertEquals(len(prdf.values()), 4)
-        self.assertAlmostEquals(
+        self.assertEqual(len(prdf.values()), 4)
+        self.assertAlmostEqual(
             prdf[('Ni', 'Al')]['distribution'][int(round(2 / 0.5))], 0.125236677)
-        self.assertAlmostEquals(
+        self.assertAlmostEqual(
             prdf[('Al', 'Ni')]['distribution'][int(round(2 / 0.5))], 0.37571003)
-        self.assertAlmostEquals(
+        self.assertAlmostEqual(
             prdf[('Al', 'Al')]['distribution'][int(round(2 / 0.5))], 0)
 
     def test_redf(self):
@@ -229,8 +227,8 @@ class StructureFeaturesTest(PymatgenTest):
             np.linalg.norm(sin_mat - np.array(mtarget)), 0.0, places=4)
         scm = SineCoulombMatrix(False)
         sin_mat = scm.featurize(self.diamond)[0]
-        self.assertEquals(sin_mat[0][0], 0)
-        self.assertEquals(sin_mat[1][1], 0)
+        self.assertEqual(sin_mat[0][0], 0)
+        self.assertEqual(sin_mat[1][1], 0)
 
     def test_orbital_field_matrix(self):
         ofm_maker = OrbitalFieldMatrix()
@@ -243,7 +241,7 @@ class StructureFeaturesTest(PymatgenTest):
         for i in range(32):
             for j in range(32):
                 if not i in [1, 3] and not j in [1, 3]:
-                    self.assertEquals(ofm[i, j], 0.0)
+                    self.assertEqual(ofm[i, j], 0.0)
         mtarget = np.matrix(mtarget)
         self.assertAlmostEqual(
             np.linalg.norm(ofm - mtarget), 0.0, places=4)